requirements.txt
# Reddit API
praw==7.7.1

# HTTP 请求
requests==2.31.0

# 定时任务
apscheduler==3.10.4

# Git 操作
gitpython==3.1.40

# 邮件发送
secure-smtplib==0.1.1

# 环境变量
python-dotenv==1.0.0

# 日期时间处理
python-dateutil==2.8.2

# 进度条
tqdm==4.66.1
